dumb way of linking projects
