+++ modules/basic-landing-page-components/package.json
@@ -2,7 +2,7 @@
   "name": "@idahoedokpayi/basic-landing-page-components",
   "version": "0.0.1",
   "description": "React Components module for basic-landing-page-components",
-  "main": "dist/index.js",
+  "main": "dist/index.js",  
   "types": "dist/src/index.d.ts",
   "scripts": {
     "test": "jest",
@@ -41,6 +41,7 @@
     "@types/webpack": "^4.41.12",
     "babel-loader": "^8.1.0",
     "babel-minify-webpack-plugin": "^0.3.1",
+    "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^3.5.3",
     "file-loader": "^6.0.0",
     "jest": "^25.5.4",
