+++ modules/basic-landing-page-components/src/ga.d.ts
@@ -1,0 +1,7 @@
+export {};
+declare global {
+    interface Window {
+      dataLayer: [];
+    }
+    const dataLayer: any[];
+  }
