+++ themes/boostrap-cover/package.json
@@ -26,7 +26,8 @@
     "gatsby": "^2.20.28",
     "gatsby-plugin-eslint": "^2.0.8",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react-dom": "^16.13.1",
+    "typescript": "^3.8.3"
   },
   "peerDependencies": {
     "gatsby": "^2.20.28",
@@ -36,6 +37,7 @@
     "popper.js": "^1.16.1"
   },
   "dependencies": {
+    "@idahoedokpayi/basic-landing-page-components": "file:../../modules/basic-landing-page-components",
     "@popperjs/core": "^2.3.3",
     "bootstrap": "^4.4.1",
     "gatsby-plugin-react-helmet": "^3.2.5",
@@ -44,7 +46,6 @@
     "jquery": "^3.5.0",
     "node-sass": "^4.13.1",
     "popper.js": "^1.16.1",
-    "react-helmet": "^6.0.0",
-    "typescript": "^3.8.3"
+    "react-helmet": "^6.0.0"
   }
 }
