--- tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist/",
-    "noImplicitAny": true,
-    "module": "commonjs",
-    "target": "es5",
-    "lib": ["ESNext", "DOM"],
-    "jsx": "react",
-    "allowJs": true,
-    "sourceMap": true,
-    "declaration": true,
-    "declarationDir": "./dist/"
-  }, 
-  "include": ["./themes/**/src/**/*"]
-}
