+++ workspace.json
@@ -3,6 +3,6 @@
       "bootstrap-cover": "themes/bootstrap-cover",
       "corporate-homepage": "themes/corporate-homepage",
       "product-landing-page":"themes/product-landing-page",
-      "basic-landing-page-components": "themes/basic-landing-page-components"
+      "@idahoedokpayi/basic-landing-page-components": "modules/basic-landing-page-components"
     }
   }
